Enable QT_NO_NARROWING_CONVERSIONS_IN_CONNECT

+++ Viewer/VideoToolBar.cpp
@@ -62,7 +62,7 @@ void VideoToolBar::setRange(int min, int max)
     m_totalTime->setText(QTime(0, 0, 0).addMSecs(max).toString(format()));
 }
 
-void VideoToolBar::setPosition(int value)
+void VideoToolBar::setPosition(qint64 value)
 {
     m_offsetSlider->setValue(value);
     m_currentOffset->setText(QTime(0, 0, 0).addMSecs(value).toString(format()));
